requirements.txt
# Core data packages - use pre-compiled versions
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0

# API clients
newsapi-python==0.2.7
google-search-results==2.4.2
tweepy==4.16.0
praw==7.8.1
requests>=2.28.0

# ML packages - specify versions that have wheels
transformers>=4.30.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
prophet>=1.1.0

# Communication
slack_sdk>=3.20.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0